+++ src/components/atoms/form/input/InputBase.vue
@@ -1,0 +1,5 @@
+<template>
+  <input
+    class="border rounded-xl p-2 w-auto py-0 px-4 leading-8 bg-background50 border-neutral200 focus:border-fontLink focus:outline-none"
+  />
+</template>