+++ src/components/atoms/form/input/InputText.vue
@@ -1,0 +1,7 @@
+<template>
+  <InputBase />
+</template>
+
+<script setup lang="ts">
+import InputBase from '@src/components/atoms/form/input/InputBase.vue';
+</script>